% Generated by build script
\name{bundlephobia_compare}
\alias{bundlephobia_compare}
\title{Compare bundle sizes of multiple npm packages}
\usage{
bundlephobia_compare()
}
\arguments{
\item{packages}{Character vector of npm package names}
}
\value{tibble: name, version, description, size, gzip,}
\description{Compare bundle sizes of multiple npm packages}
