% Generated by build script
\name{bundlephobia_history}
\alias{bundlephobia_history}
\title{Get bundle size history across versions for an npm package}
\usage{
bundlephobia_history()
}
\arguments{
\item{package}{npm package name (e.g. "react", "lodash")}
}
\value{tibble: name, version, size, gzip}
\description{Get bundle size history across versions for an npm package}
