% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundlephobia.R
\name{bundlephobia_size}
\alias{bundlephobia_size}
\title{Get npm package bundle size from Bundlephobia}
\usage{
bundlephobia_size(package)
}
\arguments{
\item{package}{npm package name (e.g. "react", "lodash", "express")}
}
\value{
tibble: name (character), version (character), description (character),
size (integer), gzip (integer), dependency_count (integer),
has_side_effects (logical), repository (character)
}
\description{
Returns the minified and gzipped bundle size for an npm package,
plus dependency count and metadata.
}
