% Generated by build script
\name{bundlephobia_size}
\alias{bundlephobia_size}
\title{Get npm package bundle size from Bundlephobia}
\usage{
bundlephobia_size()
}
\arguments{
\item{package}{npm package name (e.g. "react", "lodash", "express")}
}
\value{tibble: name (character), version (character), description (character),}
\description{Get npm package bundle size from Bundlephobia}
